datatype mismatch was creating incorrect gt_database, made the datatype of gt_points and gt_boxes consistent for custom gt_database

+++ pcdet/datasets/custom/custom_dataset.py
@@ -193,7 +193,8 @@ class CustomDataset(DatasetTemplate):
             for i in range(num_obj):
                 filename = '%s_%s_%d.bin' % (sample_idx, names[i], i)
                 filepath = database_save_path / filename
-                gt_points = points[point_indices[i] > 0]
+                gt_points = points[point_indices[i] > 0].astype(np.float32)
+                gt_boxes[i, :3] = gt_boxes[i, :3].astype(np.float32)
 
                 gt_points[:, :3] -= gt_boxes[i, :3]
                 with open(filepath, 'w') as f:
